+++ spec/models/person_spec.rb
@@ -39,38 +39,6 @@ describe Person do
       expect(people.count).to eq(1)
       expect(people.first.name).to eq('Alice Mante')
     end
-
-    context 'found in' do
-      it 'finds in which association the search term has been found' do
-        search_term = 'duckduck'
-        people = Person.search(search_term)
-        person = people[0]
-        person = Person.includes(:department, :roles, :projects, :activities,
-                                 :educations, :advanced_trainings, :expertise_topics)
-                       .find(person.id)
-        expect(person.found_in(search_term)).to eq('projects#title')
-      end
-
-      it 'finds in which person attribute the search term has been found' do
-        search_term = 'Alice'
-        people = Person.search(search_term)
-        person = people[0]
-        person = Person.includes(:department, :roles, :projects, :activities,
-                                 :educations, :advanced_trainings, :expertise_topics)
-                       .find(person.id)
-        expect(person.found_in(search_term)).to eq('name')
-      end
-
-      it 'returns nil for nonsense search term on found in' do
-        search_term = 'Alice'
-        people = Person.search(search_term)
-        person = people[0]
-        person = Person.includes(:department, :roles, :projects, :activities,
-                                 :educations, :advanced_trainings, :expertise_topics)
-                       .find(person.id)
-        expect(person.found_in('ahdkifekjsdfsakdh34ukjnfv')).to eq(nil)
-      end
-    end
   end
 
   context 'validations' do
